requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py@c055fd3
typing_extensions==4.12.2
mypy==1.10.0
PyMySQL==1.4.6
SQLAlchemy==2.0.31
